Only save number of components for origin
Do not save the entire PathBuf for origin, since we never actually uses
the PathBuf itself when iterating through files. We only use the origin
to compute the number of components in the path for the origin, and
since that value never changes, we can compute in once. This results in
both a slighlty lower memory usage and less CPU consumption.